feat(AC): moved date handler to own file
