+++ apps/frontend/src/components/molecules/Calendar-management/restriction-handlers.tsx
@@ -15,7 +15,7 @@ import { ReactNode, useState } from "react";
 import { Spinner } from "@/components/atoms/baseShadcn/spinner";
 import CalCard from "@/components/organisms/Calandar-management/temporary-card";
 
-abstract class RestrictionHandler {
+export abstract class RestrictionHandler {
   protected MyHandletypes: RestrictionTypes[];
   protected next: RestrictionHandler | null = null;
 
@@ -60,35 +60,22 @@ abstract class RestrictionHandler {
 import RestrictionTypes
 */
 
-export class DateOnlyHanlder extends RestrictionHandler {
+export class RangeDayHandler extends RestrictionHandler {
   constructor() {
-    super([
-      "SEMESTER_1_START",
-      "SEMESTER_1_END",
-      "SEMESTER_2_END",
-      "SEMESTER_2_START",
-      "UNIVERSITY_CLOSURE",
-      "PUBLIC_HOLIDAY",
-    ]);
+    super(["EXAM_PERIOD", "SUPP_WEEK", "TEST_WEEK", "HOLIDAY", "RECESS"]);
   }
   handleHtml(
     resType: SingleRestrictionResp,
     academicCalendarID: string,
   ): React.ReactNode {
     return (
-      <DateRestrictionHTML
+      <RangeDateHTML
         resType={resType}
         academicCalendarID={academicCalendarID}
       />
     );
   }
 }
-
-interface restrictionProps {
-  resType: SingleRestrictionResp;
-  academicCalendarID: string;
-}
-
 interface containerProp {
   children: ReactNode;
   type: string;
@@ -106,135 +93,9 @@ function RestrictionContainerHtml({ children, type }: containerProp) {
   );
 }
 
-function DateRestrictionHTML({
-  resType,
-  academicCalendarID,
-}: restrictionProps) {
-  const { mutate: updateMut, isPending: savePending } = useMutation(
-    UpdateRestrictionMutation,
-  );
-  const { mutate: deleteMut, isPending: deletePending } = useMutation({
-    ...DelRestrictionMut,
-  });
-
-  const [restriction, setRestriction] =
-    useState<SingleRestrictionResp>(resType);
-
-  function save() {
-    updateMut({
-      body: {
-        description: restriction.description,
-        startDate: restriction.startDate,
-        type: restriction.type,
-      },
-      paths: {
-        restrictionId: restriction.id,
-        id: academicCalendarID,
-      },
-    });
-  }
-  function deleteRes() {
-    deleteMut({
-      paths: {
-        id: academicCalendarID,
-        restrictionId: restriction.id,
-      },
-    });
-  }
-
-  return (
-    <div className="flex flex-row w-full  items-center justify-items-center gap-4 text-center ">
-      <Label className="text-sm font-medium text-[var(--text-secondary)] flex flex-col">
-        Selected date
-        <Input
-          data-testid="restriction-Date-Input"
-          type="date"
-          value={restriction.startDate}
-          onChange={(e) => {
-            if (e.target.value) {
-              {
-                setRestriction((res) => ({
-                  ...res,
-                  startDate: e.target.value,
-                }));
-              }
-            }
-          }}
-          className="h-8 rounded-md border border-[var(--border)] bg-transparent px-2 text-sm text-[var(--text-primary)] focus:outline-none focus:ring-1 focus:ring-[var(--ring)]"
-        />
-      </Label>
-      <Label className="text-sm font-medium text-[var(--text-secondary)] flex flex-col">
-        description
-        <Input
-          data-testid="restriction-dsc-Input"
-          type="text"
-          value={restriction.description}
-          onChange={(e) => {
-            if (e.target.value) {
-              {
-                setRestriction((res) => ({
-                  ...res,
-                  description: e.target.value,
-                }));
-              }
-            }
-          }}
-          className="h-8  rounded-md border border-[var(--border)] bg-transparent px-2 text-sm text-[var(--text-primary)] focus:outline-none focus:ring-1 focus:ring-[var(--ring)]"
-        />
-      </Label>
-      <Button
-        id="btn-delete-restriction"
-        type="button"
-        variant="ghost"
-        size="icon"
-        onClick={deleteRes}
-        disabled={deletePending}
-        className="h-10 w-10 flex-shrink-0 border border-[var(--error-text)] text-[var(--text-secondary)] transition-colors duration-[var(--duration-fast)] hover:border-[var(--error-text)] hover:text-[var(--error-text)] hover:bg-[var(--error-bg)]"
-      >
-        <Trash2
-          size={16}
-          strokeWidth={1.5}
-          className="text-[var(--error-text)]"
-        />
-      </Button>
-      <Button
-        id="btn-delete-restriction"
-        type="button"
-        variant="ghost"
-        size="icon"
-        onClick={save}
-        disabled={savePending}
-        className="h-10 w-10 flex-shrink-0 border border-[var(--success-text)] text-[var(--text-secondary)] transition-colors duration-[var(--duration-fast)] hover:border-[var(--success-text)] hover:text-[var(--success-text)] hover:bg-[var(--success-bg)]"
-      >
-        {savePending ? (
-          <Spinner />
-        ) : (
-          <Save
-            size={16}
-            strokeWidth={1.5}
-            className="text-[var(--success-text)]"
-          />
-        )}
-      </Button>
-    </div>
-  );
-}
-
-export class RangeDayHandler extends RestrictionHandler {
-  constructor() {
-    super(["EXAM_PERIOD", "SUPP_WEEK", "TEST_WEEK", "HOLIDAY", "RECESS"]);
-  }
-  handleHtml(
-    resType: SingleRestrictionResp,
-    academicCalendarID: string,
-  ): React.ReactNode {
-    return (
-      <RangeDateHTML
-        resType={resType}
-        academicCalendarID={academicCalendarID}
-      />
-    );
-  }
+export interface restrictionProps {
+  resType: SingleRestrictionResp;
+  academicCalendarID: string;
 }
 
 function RangeDateHTML({ resType, academicCalendarID }: restrictionProps) {
